requirements.txt
beautifulsoup4
psycopg
PyYAML
opendartreader
finance-datareader
numpy
pandas
matplotlib
seaborn
jupyterlab
rdflib
pyvis
streamlit
spacy[transformers]
spacy-streamlit